Remove invalid java invocation arguments
The UseConcMarkSweepGC, CMSIncrementalPacing, and AggressiveOpts
arguments are no longer supported in current versions of Java.
Java fails to start with them present.

Additionally, having less specific default invocation options
allows supporting more versions of Java and makes for less
maintenance into the future as options are deprecated.

Invocation options can be added to individual server configs, or
to the global config, if the user wants to change the defaults.

Fixes issue #402