Update wl, bl, and op commands to not edit files
This commit updates the wl, bl, and op commands to no longer edit
their corresponding server files.  msm still edits config settings
in files, but relies on minecraft server to manage it's own files,
passing the commands to the running server for changes.

Minecraft server moved to json in version 1.7.2 which made
directly editing txt files no longer work.

This change follows the pattern of the other passthrough commands
that if the server isn't running, returns error_exit informing the
user the server is not running.

Fixes issue #404